Modified Signup Page
Changed signup page to work with the logic on the mock-form page that previously existed. The original code that was there has been commented at the bottom.
